refactor(Modularise Code): Move App.jsx logic to own file and use it in Home.jsx
