+++ src/pages/Home.jsx
@@ -1,0 +1,7 @@
+import WasmImageProcessor from '@components/WasmImageProcessor';
+
+const Home = () => {
+	return <WasmImageProcessor />
+}
+
+export default Home;
